% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MouseImmuneReference-package.R
\docType{package}
\name{MouseImmuneReference}
\alias{MouseImmuneReference}
\alias{MouseImmuneReference-package}
\alias{MouseImmuneReference.run}
\title{MouseImmuneReference}
\usage{
MouseImmuneReference.run(launch.browser = TRUE, ...)
}
\description{
MouseImmuneReference
}

